Refactor: users 엔드포인트 기능별로 파일 구분
에러핸들링이 필요합니다
